Search in several pages and returned results
Fixed the algorithm to recover the number of results returned by google scholar. It now is able to parse the number of results when these are well defined by google and when it is an approximation. Also it is able of obtaining the number of results regardless of the page we are.
Fixed the SearchScholarQuery and ClusterScholarQuery in order to obtain all the results returned by google scholar. It is limited to 1000 results (google's limit). The count option limits the number of final results returned by the query.